Tighten bound on optparse-applicative in library
There were two dependencies on optparse-applicative - one from the library
and one from the test suite. The library's bounds were >= 0.11 && < 0.14 and
the test suite's were >= 0.11 && < 0.13. Since cabal picks one version of a
dependency per package, if tests were enabled you got 0.12, and if not you
got 0.13. The build fails on 0.13 because they removed (<>) from
Options.Applicative. This commit tightens the bound on the library component.